refactor cronjobs to modules inyecting dependency to app
